refactor: 共通基盤を構築しデータ層を分離 (#4)
- ディレクトリ skeleton: features/entities/shared/mocks
- テスト環境: vitest + happy-dom + @testing-library/react
- shared/lib/time.js: formatDate/timeToMin/fmtDuration/fmtMin (テスト付き)
- shared/lib/markdown.jsx: renderMarkdown (テスト付き)
- entities/project/projects.js: PROJECTS/projectOrder
- mocks/*.js: events/tasks/history/today
- prototype.jsx: データ定義とユーティリティをimportに置換
