+++ src/mocks/history.js
@@ -1,0 +1,79 @@
+export const historyData = [
+  {
+    id: "h1",
+    project: "career",
+    title: "転職ドラフト応募完了",
+    date: "2026-04-05",
+    done: true,
+  },
+  {
+    id: "h2",
+    project: "slo",
+    title: "SLI候補の洗い出し",
+    date: "2026-04-05",
+    done: true,
+  },
+  {
+    id: "h3",
+    project: "loadtest",
+    title: "Locust環境セットアップ",
+    date: "2026-04-06",
+    done: true,
+  },
+  {
+    id: "h4",
+    project: "tasuki",
+    title: "php-parser PoC完了",
+    date: "2026-04-06",
+    done: true,
+  },
+  {
+    id: "h5",
+    project: "career",
+    title: "Finatext オファー検討",
+    date: "2026-04-07",
+    done: true,
+  },
+  {
+    id: "h6",
+    project: "slo",
+    title: "New Relicアラート設定",
+    date: "2026-04-07",
+    done: true,
+  },
+  {
+    id: "h7",
+    project: "loadtest",
+    title: "WireMock導入調査",
+    date: "2026-04-08",
+    done: true,
+  },
+  {
+    id: "h8",
+    project: "career",
+    title: "ULS企業研究メモ作成",
+    date: "2026-04-09",
+    done: true,
+  },
+  {
+    id: "h9",
+    project: "tasuki",
+    title: "Terminal埋め込みPoC",
+    date: "2026-04-09",
+    done: true,
+  },
+  {
+    id: "h10",
+    project: "slo",
+    title: "エラーバジェットポリシー草案",
+    date: "2026-04-10",
+    done: true,
+  },
+  {
+    id: "h11",
+    project: "loadtest",
+    title: "ECS Fargate構成設計",
+    date: "2026-04-10",
+    done: true,
+  },
+];
