+++ src/mocks/today.js
@@ -1,0 +1,1 @@
+export const TODAY = "2026-04-11";
