+++ src/prototype.jsx
@@ -1,423 +1,11 @@
 import { useState, useCallback, useRef, useMemo } from "react";
-
-// ─── Data ───────────────────────────────────────────────────────────
-const PROJECTS = {
-  career: { name: "転職活動", color: "#E85D04" },
-  loadtest: { name: "負荷試験", color: "#0096C7" },
-  slo: { name: "SLO推進", color: "#2D9F45" },
-  tasuki: { name: "Tasuki", color: "#9B5DE5" },
-};
-const projectOrder = ["career", "loadtest", "slo", "tasuki"];
-const TODAY = "2026-04-11";
-
-const initialEvents = [
-  {
-    id: "e2",
-    title: "SLOレビューMTG",
-    time: "09:30",
-    endTime: "10:30",
-    date: TODAY,
-    project: "slo",
-    meetUrl: "https://meet.google.com/abc-defg-hij",
-    attachments: ["SLI定義書_v2.pdf", "エラーバジェットポリシー草案.docx"],
-    description:
-      "## アジェンダ\n\n1. SLI定義の最終確認\n2. エラーバジェットポリシーのレビュー\n3. New Relicダッシュボードのデモ\n\n## 参加者\n\n- 田中PM\n- インフラチーム\n- 自分",
-  },
-  {
-    id: "e1",
-    title: "デイリースタンドアップ",
-    time: "11:00",
-    endTime: "11:15",
-    date: TODAY,
-    description:
-      "チーム全体の進捗共有\n- 各自の今日のタスク確認\n- ブロッカーの共有",
-  },
-  {
-    id: "e3",
-    title: "Dirbato最終面接",
-    time: "14:00",
-    endTime: "15:00",
-    date: TODAY,
-    project: "career",
-    meetUrl: "https://zoom.us/j/123456789",
-    attachments: ["職務経歴書_最終版.pdf"],
-    description:
-      "## 面接情報\n\n- 面接官: 執行役員 佐藤氏\n- 形式: オンライン (Zoom)\n\n## 準備\n\n- 志望動機の最終整理\n- 逆質問3つ用意\n- `技術力 × ビジネス理解` の軸で話す",
-  },
-  {
-    id: "e4",
-    title: "1on1 with マネージャー",
-    time: "17:00",
-    endTime: "17:30",
-    date: TODAY,
-    meetUrl: "https://meet.google.com/xyz-uvwx-rst",
-    description:
-      "- 今週の振り返り\n- 来週の優先順位確認\n- キャリアの相談（転職活動の進捗共有）",
-  },
-  {
-    id: "e5",
-    title: "もくもく会",
-    time: "21:00",
-    endTime: "23:00",
-    date: TODAY,
-    meetUrl: "https://meet.google.com/moku-moku-kai",
-    description: "オンラインもくもく会\n\n- 各自作業\n- 30分ごとに進捗共有",
-  },
-];
-
-const initialTasks = [
-  {
-    id: "t1",
-    project: "career",
-    title: "面接対策：志望動機の最終整理",
-    size: "M",
-    done: false,
-    dependsOn: "e3",
-    body: "## やること\n\n- **志望動機**を3パターン用意\n- 技術的な強みの整理\n  - DDD / Clean Architecture\n  - SLI/SLO導入経験\n- 逆質問リストの準備\n\n## 参考\n\n`転職ドラフト`のフィードバックを確認\n\n> 上流工程への関与意欲が伝わる内容にすること",
-  },
-  {
-    id: "t2",
-    project: "slo",
-    title: "SLI定義ドキュメント更新",
-    size: "M",
-    done: false,
-    dependsOn: "e2",
-    body: "## 対象\n\nWeb Cart フロントエンドの SLI\n\n## 更新内容\n\n- Availability SLI: `成功リクエスト / 全リクエスト`\n- Latency SLI: `p99 < 500ms`\n- 計測ポイントをNew Relicの`Transaction`に合わせる",
-  },
-  {
-    id: "t3",
-    project: "loadtest",
-    title: "WireMock stub定義作成",
-    size: "M",
-    done: false,
-    dependsOn: null,
-    body: "chaos test用のstub定義ファイルを作成する\n\n### エンドポイント\n\n1. `/api/v1/orders` — 正常レスポンス\n2. `/api/v1/orders` — 429 レスポンス (rate limit)\n3. `/api/v1/payments` — 500ms遅延",
-  },
-  {
-    id: "t4",
-    project: "career",
-    title: "職務経歴書PDF最終版を送付",
-    size: "S",
-    done: false,
-    dependsOn: null,
-    body: "最終版PDFを浅野さんにメール送付\n\n確認ポイント:\n- 誤字脱字チェック済み\n- BASE在籍期間の記載が正確か",
-  },
-  {
-    id: "t5",
-    project: "tasuki",
-    title: "AnalyzerContract trait設計",
-    size: "L",
-    done: false,
-    dependsOn: null,
-    body: "## 目的\n\ncode-inspector用の抽象化層を設計する\n\n## 設計メモ\n\n```rust\ntrait AnalyzerContract {\n    fn analyze(&self, input: &SourceFile) -> AnalysisResult;\n    fn supports(&self, file_type: &FileType) -> bool;\n}\n```\n\n- php-parser と tree-sitter の両方に対応\n- call chain 解析は別traitに分離",
-  },
-  {
-    id: "t6",
-    project: "loadtest",
-    title: "Locustシナリオ実装",
-    size: "L",
-    done: false,
-    dependsOn: null,
-    body: "ピーク負荷パターンのシナリオを実装\n\n## 要件\n\n- 通常: 100 RPS\n- ピーク: 500 RPS (10分間)\n- ramp-up: 5分\n\n## メモ\n\n分散モードで ECS 上に展開予定",
-  },
-];
-
-const historyData = [
-  {
-    id: "h1",
-    project: "career",
-    title: "転職ドラフト応募完了",
-    date: "2026-04-05",
-    done: true,
-  },
-  {
-    id: "h2",
-    project: "slo",
-    title: "SLI候補の洗い出し",
-    date: "2026-04-05",
-    done: true,
-  },
-  {
-    id: "h3",
-    project: "loadtest",
-    title: "Locust環境セットアップ",
-    date: "2026-04-06",
-    done: true,
-  },
-  {
-    id: "h4",
-    project: "tasuki",
-    title: "php-parser PoC完了",
-    date: "2026-04-06",
-    done: true,
-  },
-  {
-    id: "h5",
-    project: "career",
-    title: "Finatext オファー検討",
-    date: "2026-04-07",
-    done: true,
-  },
-  {
-    id: "h6",
-    project: "slo",
-    title: "New Relicアラート設定",
-    date: "2026-04-07",
-    done: true,
-  },
-  {
-    id: "h7",
-    project: "loadtest",
-    title: "WireMock導入調査",
-    date: "2026-04-08",
-    done: true,
-  },
-  {
-    id: "h8",
-    project: "career",
-    title: "ULS企業研究メモ作成",
-    date: "2026-04-09",
-    done: true,
-  },
-  {
-    id: "h9",
-    project: "tasuki",
-    title: "Terminal埋め込みPoC",
-    date: "2026-04-09",
-    done: true,
-  },
-  {
-    id: "h10",
-    project: "slo",
-    title: "エラーバジェットポリシー草案",
-    date: "2026-04-10",
-    done: true,
-  },
-  {
-    id: "h11",
-    project: "loadtest",
-    title: "ECS Fargate構成設計",
-    date: "2026-04-10",
-    done: true,
-  },
-];
-
-function formatDate(ds) {
-  const d = new Date(ds + "T00:00:00");
-  const days = ["日", "月", "火", "水", "木", "金", "土"];
-  return `${d.getMonth() + 1}/${d.getDate()} (${days[d.getDay()]})`;
-}
-function timeToMin(t) {
-  const [h, m] = t.split(":").map(Number);
-  return h * 60 + m;
-}
-
-// ─── Simple Markdown Renderer ───────────────────────────────────────
-function renderMarkdown(md) {
-  if (!md) return null;
-  const lines = md.split("\n");
-  const elements = [];
-  let i = 0;
-  let key = 0;
-
-  const inlineStyle = (text) => {
-    // code
-    let result = text.replace(
-      /`([^`]+)`/g,
-      '<code style="background:#27272a;padding:1px 5px;border-radius:3px;font-size:0.9em;color:#a1a1aa">$1</code>',
-    );
-    // bold
-    result = result.replace(
-      /\*\*([^*]+)\*\*/g,
-      '<strong style="color:#e4e4e7;font-weight:600">$1</strong>',
-    );
-    // italic
-    result = result.replace(/\*([^*]+)\*/g, "<em>$1</em>");
-    return result;
-  };
-
-  while (i < lines.length) {
-    const line = lines[i];
-
-    // Empty line
-    if (line.trim() === "") {
-      i++;
-      continue;
-    }
-
-    // Headings
-    const hMatch = line.match(/^(#{1,3})\s+(.+)/);
-    if (hMatch) {
-      const level = hMatch[1].length;
-      const sizes = { 1: 15, 2: 13, 3: 12 };
-      elements.push(
-        <div
-          key={key++}
-          style={{
-            fontSize: sizes[level],
-            fontWeight: 600,
-            color: "#e4e4e7",
-            marginTop: elements.length ? 14 : 0,
-            marginBottom: 6,
-            fontFamily: "'Noto Sans JP', sans-serif",
-          }}
-          dangerouslySetInnerHTML={{ __html: inlineStyle(hMatch[2]) }}
-        />,
-      );
-      i++;
-      continue;
-    }
-
-    // Blockquote
-    if (line.startsWith("> ")) {
-      elements.push(
-        <div
-          key={key++}
-          style={{
-            borderLeft: "2px solid #3f3f46",
-            paddingLeft: 10,
-            margin: "6px 0",
-            color: "#71717a",
-            fontSize: 12,
-            fontStyle: "italic",
-            fontFamily: "'Noto Sans JP', sans-serif",
-          }}
-          dangerouslySetInnerHTML={{ __html: inlineStyle(line.slice(2)) }}
-        />,
-      );
-      i++;
-      continue;
-    }
-
-    // Code block
-    if (line.startsWith("```")) {
-      const lang = line.slice(3).trim();
-      const codeLines = [];
-      i++;
-      while (i < lines.length && !lines[i].startsWith("```")) {
-        codeLines.push(lines[i]);
-        i++;
-      }
-      i++; // skip closing ```
-      elements.push(
-        <pre
-          key={key++}
-          style={{
-            background: "#18181b",
-            border: "1px solid #27272a",
-            borderRadius: 6,
-            padding: "10px 12px",
-            margin: "6px 0",
-            fontSize: 11,
-            color: "#a1a1aa",
-            overflow: "auto",
-            lineHeight: 1.5,
-            whiteSpace: "pre",
-          }}
-        >
-          {lang && (
-            <div style={{ fontSize: 9, color: "#52525b", marginBottom: 4 }}>
-              {lang}
-            </div>
-          )}
-          {codeLines.join("\n")}
-        </pre>,
-      );
-      continue;
-    }
-
-    // Unordered list
-    if (line.match(/^[-*]\s/)) {
-      const items = [];
-      while (i < lines.length && lines[i].match(/^(\s*)[-*]\s/)) {
-        const m = lines[i].match(/^(\s*)[-*]\s+(.*)/);
-        const indent = m[1].length > 0 ? 1 : 0;
-        items.push({ text: m[2], indent });
-        i++;
-      }
-      elements.push(
-        <div key={key++} style={{ margin: "4px 0" }}>
-          {items.map((it, j) => (
-            <div
-              key={j}
-              style={{
-                display: "flex",
-                gap: 6,
-                marginLeft: it.indent * 16,
-                padding: "2px 0",
-                fontSize: 12,
-                color: "#a1a1aa",
-                fontFamily: "'Noto Sans JP', sans-serif",
-              }}
-            >
-              <span style={{ color: "#52525b", flexShrink: 0 }}>•</span>
-              <span
-                dangerouslySetInnerHTML={{ __html: inlineStyle(it.text) }}
-              />
-            </div>
-          ))}
-        </div>,
-      );
-      continue;
-    }
-
-    // Ordered list
-    if (line.match(/^\d+\.\s/)) {
-      const items = [];
-      while (i < lines.length && lines[i].match(/^\d+\.\s/)) {
-        items.push(lines[i].replace(/^\d+\.\s+/, ""));
-        i++;
-      }
-      elements.push(
-        <div key={key++} style={{ margin: "4px 0" }}>
-          {items.map((it, j) => (
-            <div
-              key={j}
-              style={{
-                display: "flex",
-                gap: 6,
-                padding: "2px 0",
-                fontSize: 12,
-                color: "#a1a1aa",
-                fontFamily: "'Noto Sans JP', sans-serif",
-              }}
-            >
-              <span
-                style={{
-                  color: "#52525b",
-                  flexShrink: 0,
-                  minWidth: 16,
-                  textAlign: "right",
-                }}
-              >
-                {j + 1}.
-              </span>
-              <span dangerouslySetInnerHTML={{ __html: inlineStyle(it) }} />
-            </div>
-          ))}
-        </div>,
-      );
-      continue;
-    }
-
-    // Paragraph
-    elements.push(
-      <p
-        key={key++}
-        style={{
-          fontSize: 12,
-          color: "#a1a1aa",
-          lineHeight: 1.7,
-          margin: "4px 0",
-          fontFamily: "'Noto Sans JP', sans-serif",
-        }}
-        dangerouslySetInnerHTML={{ __html: inlineStyle(line) }}
-      />,
-    );
-    i++;
-  }
-  return elements;
-}
+import { PROJECTS, projectOrder } from "./entities/project/projects.js";
+import { TODAY } from "./mocks/today.js";
+import { initialEvents } from "./mocks/events.js";
+import { initialTasks } from "./mocks/tasks.js";
+import { historyData } from "./mocks/history.js";
+import { formatDate, timeToMin, fmtDuration, fmtMin } from "./shared/lib/time.js";
+import { renderMarkdown } from "./shared/lib/markdown.jsx";
 
 // ─── Detail Panel ───────────────────────────────────────────────────
 function DetailPanel({ task, events, onClose, onUpdate, onToggleDone }) {
@@ -714,12 +302,6 @@ function EventDetailPanel({ event, onClose }) {
   const evStart = timeToMin(event.time);
   const evEnd = timeToMin(event.endTime);
   const duration = evEnd - evStart;
-  const fmtDur = (m) => {
-    if (m >= 60)
-      return `${Math.floor(m / 60)}h${m % 60 > 0 ? String(m % 60).padStart(2, "0") + "m" : ""}`;
-    return `${m}m`;
-  };
-
   const meetLabel = event.meetUrl?.includes("zoom")
     ? "Zoom"
     : event.meetUrl?.includes("meet.google")
@@ -816,7 +398,7 @@ function EventDetailPanel({ event, onClose }) {
                 fontVariantNumeric: "tabular-nums",
               }}
             >
-              {event.time}–{event.endTime} ({fmtDur(duration)})
+              {event.time}–{event.endTime} ({fmtDuration(duration)})
             </span>
           </div>
           <h2
@@ -1266,14 +848,6 @@ function StackView({
   const minutesUntilNext = nextEvent
     ? timeToMin(nextEvent.time) - nowMin
     : DAY_END - nowMin;
-
-  const fmtMin = (m) =>
-    `${Math.floor(m / 60)}:${String(m % 60).padStart(2, "0")}`;
-  const fmtDuration = (m) => {
-    if (m >= 60)
-      return `${Math.floor(m / 60)}h${m % 60 > 0 ? String(m % 60).padStart(2, "0") + "m" : ""}`;
-    return `${m}m`;
-  };
 
   return (
     <div style={{ padding: "0 0 100px" }}>
